requirements.txt
# Core GUI and image processing
tkinter
Pillow>=9.0.0
pyperclip>=1.8.2

# Drag and drop functionality
tkinterdnd2>=0.3.0

# Standard library modules (included with Python, listed for reference)
# json
# os
# sys
# glob
# threading
# typing
